requirements.txt
blinker==1.8.2
boto3==1.35.20
botocore==1.35.20
click==8.1.7
et-xmlfile==1.1.0
ezdxf==1.3.2
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Levenshtein==0.25.1
MarkupSafe==2.1.5
numpy==2.0.1
openpyxl==3.1.5
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
rapidfuzz==3.9.5
s3transfer==0.10.2
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.3
XlsxWriter==3.2.0
